Completely rework the organization of the code
This all started when I realized that in order to make n.FindValue()
API as easy to use as possible, it was going to require separate
namespaces.

Then on, it was basically just solving a really big puzzle over
how to break circular dependencies while keeping the API compatible.

Details are a little shaky.